fix: make the preview match the export at the corners, the fit and the badge
The wrapper drew a rounded plate while the canvas painted a square bitmap over
it, so the preview showed square corners where the export clips the document to
the same rounded rect. The canvas clips itself while drawing — not through a
layer corner radius, which the compositor applies and every direct render path
would miss.

A preset change resized the document without refitting, so adding a frame while
fitted pushed the new padding out of the viewport and removing one left the
document under-zoomed. The refit is scheduled after the layout pass, as an
image change already does, and the fit seam measures the outer size instead of
the image, which could call a document fitted while its backdrop was cut off.

gradientAxis truncated its size to whole units, so a 121x81 at scale 2 framed
to 140.5 points on screen but drew its gradient along a 140-point axis — tilted
away from the export's. It takes the size as it is. The badge stopped
truncating for the same reason, and now reports the size of what Save would
write, which follows a pending crop; the wrapper and fit keep describing the
live document being cropped.

Gate fixes: the export sentinel is asymmetric and sits inside the corner
radius, because the old marker sat in the very corner the export legitimately
clips away and would have failed correct code; the corner is pinned as
backdrop rather than document; the live wrapper is rasterized with its subview
and asked the same questions as the export, so a preview that drew nothing can
no longer pass on geometry alone; an odd-pixel Retina case covers the half
point; fit is left to happen on its own; and the follow-the-document gate puts
each mutation in its own undo group, since NSUndoManager coalesces by event and
one undo would otherwise revert both the crop and the backdrop.

+++ Sources/Snippr/SliceBContract.swift
@@ -908,8 +908,7 @@ enum SliceBBackdrop {
             colorsSpace: CGColorSpace(name: CGColorSpace.sRGB)!,
             colors: gradient(for: preset) as CFArray, locations: [0, 1])
         else { return false }
-        let axis = gradientAxis(
-            width: Int(size.width), height: Int(size.height))
+        let axis = gradientAxis(width: size.width, height: size.height)
         ctx.drawLinearGradient(
             fill, start: axis.start, end: axis.end, options: [])
         let scale = max(1, pixelScale)
@@ -991,10 +990,13 @@ enum SliceBBackdrop {
 
     /// Gradient axis for a frame of this size, shared by `compose` and by
     /// anything reproducing the background.
+    /// Takes the size as it IS, not truncated to whole units: a 121x81 @2
+    /// document frames to 140.5 x 120.5 pt on screen, and rounding that down
+    /// tilts the on-screen gradient away from the exported one.
     static func gradientAxis(
-        width: Int, height: Int
+        width: CGFloat, height: CGFloat
     ) -> (start: CGPoint, end: CGPoint) {
-        (CGPoint(x: 0, y: CGFloat(height)), CGPoint(x: CGFloat(width), y: 0))
+        (CGPoint(x: 0, y: height), CGPoint(x: width, y: 0))
     }
 
     private static func gradient(for preset: BackdropPreset) -> [CGColor] {
